+++ src/main/java/SQL/models/Customer.java
@@ -1,4 +1,4 @@
-package sql.models;
+package SQL.models;
 
 public class Customer {
     private Name name;
