+++ src/test/java/load/sqldatabase/CustomerDAOTest.java
@@ -5,7 +5,7 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.BeforeEach;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import sql.dao.CustomerDAO;
+import SQL.DAO.CustomerDAO;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
