+++ src/test/java/mongoReader/MongoConnectorTest.java
@@ -1,11 +1,10 @@
 package mongoReader;
 
-import Invoices.Invoice;
-import Invoices.InvoiceDAO;
+import invoices.Invoice;
+import invoices.DAO.InvoiceDAO;
 import com.mongodb.client.MongoCollection;
 import org.bson.Document;
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 
 import java.util.ArrayList;
@@ -43,7 +42,6 @@ class MongoConnectorTest {
 
     }
 
-    @Disabled
     @Test
     void testGetInvoiceInfo() {
         ArrayList<Invoice> invoices = invoiceDAO.getInvoices();
